# apps/backend-api/src/ibreeze_backend/auth/router.py
"""Authentication HTTP routes defined by design section G.11."""

import uuid

from fastapi import APIRouter, Depends, Request, Response, status
from sqlalchemy.ext.asyncio import AsyncSession

from ibreeze_backend.api.errors import raise_problem
from ibreeze_backend.api.schemas import success_response
from ibreeze_backend.auth.schemas import (
    AuthKeysResponse,
    ChangePasswordRequest,
    LoginRequest,
    RefreshRequest,
    RegisterRequest,
    RegisterResponse,
    SessionResponse,
)
from ibreeze_backend.auth.service import (
    ADMIN_AUDIENCE,
    APP_AUDIENCE,
    SESSION_SECONDS,
    admin_login,
    change_password,
    get_auth_keys,
    login,
    logout,
    logout_all,
    refresh_tokens,
    register,
    user_payload,
    verify_access_token,
)
from ibreeze_backend.db.session import get_db_session
from ibreeze_backend.dependencies import get_current_user
from ibreeze_backend.models.user import User
from ibreeze_backend.observability.logging_config import get_logger

logger = get_logger("ibreeze.auth")

router = APIRouter(prefix="/api/v1/auth", tags=["auth"])
admin_router = APIRouter(prefix="/admin/api/v1/auth", tags=["admin-auth"])

_ADMIN_REFRESH_COOKIE = "ibreeze_admin_refresh"


def _request_id(request: Request) -> str:
    return str(request.state.request_id)


def _set_no_store(response: Response) -> None:
    response.headers["Cache-Control"] = "no-store"
    response.headers["Pragma"] = "no-cache"


def _extract_access_token(request: Request) -> str:
    authorization = request.headers.get("authorization", "")
    if not authorization.startswith("Bearer "):
        raise_problem(401, "AUTH_MISSING_TOKEN", "Missing authorization token")
    return authorization[7:]


def _session_id(request: Request, audience: str) -> uuid.UUID:
    payload = verify_access_token(_extract_access_token(request), audience)
    if payload is None:
        raise_problem(401, "AUTH_TOKEN_EXPIRED", "Invalid or expired token")
    assert payload is not None
    return uuid.UUID(payload["sid"])


def _set_admin_refresh_cookie(response: Response, refresh_token: str) -> None:
    response.set_cookie(
        key=_ADMIN_REFRESH_COOKIE,
        value=refresh_token,
        max_age=SESSION_SECONDS,
        secure=True,
        httponly=True,
        samesite="strict",
        path="/admin/api/v1/auth",
    )


def _clear_admin_refresh_cookie(response: Response) -> None:
    response.delete_cookie(
        key=_ADMIN_REFRESH_COOKIE,
        secure=True,
        httponly=True,
        samesite="strict",
        path="/admin/api/v1/auth",
    )


def _admin_public_session(result: dict[str, object]) -> dict[str, object]:
    return {key: value for key, value in result.items() if key not in {"refresh_token", "refresh_token_expires_in"}}


@router.post(
    "/register",
    response_model=RegisterResponse,
    status_code=status.HTTP_201_CREATED,
)
async def register_endpoint(
    body: RegisterRequest,
    request: Request,
    db: AsyncSession = Depends(get_db_session),
) -> dict[str, object]:
    logger.info("register.start")
    try:
        user = await register(db, str(body.email), body.password)
        logger.info("register.completed", extra={"user_id": str(user.id)})
    except ValueError as exc:
        logger.error("register.failed", extra={"error": str(exc)})
        raise_problem(409, "AUTH_EMAIL_EXISTS", str(exc))
    return success_response(
        {"user": user_payload(user)},
        _request_id(request),
    )


@router.post(
    "/login",
    response_model=SessionResponse,
    response_model_exclude_none=True,
)
async def login_endpoint(
    body: LoginRequest,
    request: Request,
    response: Response,
    db: AsyncSession = Depends(get_db_session),
) -> dict[str, object]:
    logger.info("login.start")
    try:
        result = await login(
            db,
            body.identifier,
            body.password,
            APP_AUDIENCE,
            body.device_id,
        )
        logger.info("login.completed")
    except ValueError as exc:
        logger.error("login.failed", extra={"error": str(exc)})
        raise_problem(401, "AUTH_INVALID_CREDENTIALS", str(exc))
    _set_no_store(response)
    return success_response(result, _request_id(request))


@router.post(
    "/refresh",
    response_model=SessionResponse,
    response_model_exclude_none=True,
)
async def refresh_endpoint(
    body: RefreshRequest,
    request: Request,
    response: Response,
    db: AsyncSession = Depends(get_db_session),
) -> dict[str, object]:
    logger.info("refresh.start")
    try:
        result = await refresh_tokens(db, body.refresh_token, APP_AUDIENCE)
        logger.info("refresh.completed")
    except ValueError as exc:
        logger.error("refresh.failed", extra={"error": str(exc)})
        code = "AUTH_TOKEN_REPLAY" if "replay" in str(exc) else "AUTH_TOKEN_INVALID"
        raise_problem(401, code, str(exc))
    _set_no_store(response)
    return success_response(result, _request_id(request))


@router.post("/logout", status_code=status.HTTP_204_NO_CONTENT)
async def logout_endpoint(
    request: Request,
    db: AsyncSession = Depends(get_db_session),
    _current_user: User = Depends(get_current_user),
) -> None:
    logger.info("logout.start")
    await logout(db, _extract_access_token(request), APP_AUDIENCE)
    logger.info("logout.completed")


@router.post("/logout-all", status_code=status.HTTP_204_NO_CONTENT)
async def logout_all_endpoint(
    db: AsyncSession = Depends(get_db_session),
    current_user: User = Depends(get_current_user),
) -> None:
    logger.info("logout_all.start", extra={"user_id": str(current_user.id)})
    await logout_all(db, current_user.id)
    logger.info("logout_all.completed", extra={"user_id": str(current_user.id)})


@router.post(
    "/change-password",
    response_model=SessionResponse,
    response_model_exclude_none=True,
)
async def change_password_endpoint(
    body: ChangePasswordRequest,
    request: Request,
    response: Response,
    db: AsyncSession = Depends(get_db_session),
    current_user: User = Depends(get_current_user),
) -> dict[str, object]:
    logger.info("change_password.start", extra={"user_id": str(current_user.id)})
    try:
        result = await change_password(
            db,
            current_user.id,
            body.current_password,
            body.new_password,
            _session_id(request, APP_AUDIENCE),
            APP_AUDIENCE,
        )
        logger.info("change_password.completed", extra={"user_id": str(current_user.id)})
    except ValueError as exc:
        logger.error("change_password.failed", extra={"user_id": str(current_user.id), "error": str(exc)})
        code = "AUTH_INVALID_SESSION" if "session" in str(exc) else "AUTH_INVALID_CREDENTIALS"
        raise_problem(400, code, str(exc))
    _set_no_store(response)
    return success_response(result, _request_id(request))


@router.get("/keys", response_model=AuthKeysResponse)
async def get_keys_endpoint(request: Request) -> dict[str, object]:
    logger.info("get_keys.start")
    result = success_response(get_auth_keys(), _request_id(request))
    logger.info("get_keys.completed")
    return result


@admin_router.post(
    "/login",
    response_model=SessionResponse,
    response_model_exclude_none=True,
)
async def admin_login_endpoint(
    body: LoginRequest,
    request: Request,
    response: Response,
    db: AsyncSession = Depends(get_db_session),
) -> dict[str, object]:
    logger.info("admin_login.start")
    try:
        result = await admin_login(
            db,
            body.identifier,
            body.password,
            body.device_id,
        )
        logger.info("admin_login.completed")
    except ValueError as exc:
        logger.error("admin_login.failed", extra={"error": str(exc)})
        raise_problem(401, "AUTH_INVALID_CREDENTIALS", str(exc))
    _set_admin_refresh_cookie(response, str(result["refresh_token"]))
    _set_no_store(response)
    return success_response(_admin_public_session(result), _request_id(request))


@admin_router.post(
    "/refresh",
    response_model=SessionResponse,
    response_model_exclude_none=True,
)
async def admin_refresh_endpoint(
    request: Request,
    response: Response,
    db: AsyncSession = Depends(get_db_session),
) -> dict[str, object]:
    logger.info("admin_refresh.start")
    refresh_token = request.cookies.get(_ADMIN_REFRESH_COOKIE)
    if refresh_token is None:
        logger.error("admin_refresh.failed", extra={"error": "missing_refresh_token"})
        raise_problem(401, "AUTH_MISSING_TOKEN", "Missing refresh token")
    assert refresh_token is not None
    try:
        result = await refresh_tokens(db, refresh_token, ADMIN_AUDIENCE)
        logger.info("admin_refresh.completed")
    except ValueError as exc:
        logger.error("admin_refresh.failed", extra={"error": str(exc)})
        _clear_admin_refresh_cookie(response)
        code = "AUTH_TOKEN_REPLAY" if "replay" in str(exc) else "AUTH_TOKEN_INVALID"
        raise_problem(401, code, str(exc))
    _set_admin_refresh_cookie(response, str(result["refresh_token"]))
    _set_no_store(response)
    return success_response(_admin_public_session(result), _request_id(request))


@admin_router.post("/logout", status_code=status.HTTP_204_NO_CONTENT)
async def admin_logout_endpoint(
    request: Request,
    response: Response,
    db: AsyncSession = Depends(get_db_session),
    _current_user: User = Depends(get_current_user),
) -> None:
    logger.info("admin_logout.start")
    await logout(db, _extract_access_token(request), ADMIN_AUDIENCE)
    _clear_admin_refresh_cookie(response)
    logger.info("admin_logout.completed")


@admin_router.post(
    "/change-password",
    response_model=SessionResponse,
    response_model_exclude_none=True,
)
async def admin_change_password_endpoint(
    body: ChangePasswordRequest,
    request: Request,
    response: Response,
    db: AsyncSession = Depends(get_db_session),
    current_user: User = Depends(get_current_user),
) -> dict[str, object]:
    logger.info("admin_change_password.start", extra={"user_id": str(current_user.id)})
    try:
        result = await change_password(
            db,
            current_user.id,
            body.current_password,
            body.new_password,
            _session_id(request, ADMIN_AUDIENCE),
            ADMIN_AUDIENCE,
        )
        logger.info("admin_change_password.completed", extra={"user_id": str(current_user.id)})
    except ValueError as exc:
        logger.error("admin_change_password.failed", extra={"user_id": str(current_user.id), "error": str(exc)})
        code = "AUTH_INVALID_SESSION" if "session" in str(exc) else "AUTH_INVALID_CREDENTIALS"
        raise_problem(400, code, str(exc))
    _set_admin_refresh_cookie(response, str(result["refresh_token"]))
    _set_no_store(response)
    return success_response(_admin_public_session(result), _request_id(request))